backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
google-generativeai
supabase